# db: Optimize authzFilter code path