--- a/cmd/frontend/authz/common_test.go
+++ b/cmd/frontend/authz/common_test.go
@@ -1,1 +1,0 @@
-package authz