--- a/cmd/frontend/authz/perms.go
+++ b/cmd/frontend/authz/perms.go
@@ -1,8 +1,6 @@
 package authz
 
 import (
-	"strings"
-
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
 )
 
@@ -18,33 +16,27 @@ type Perms uint32
 
 // Perm constants.
 const (
-	Read Perms = 1 << (32 - 1 - iota)
-	Write
 	None Perms = 0
+	Read Perms = 1 << iota
+	Write
 )
 
 // Include is a convenience method to test if Perms
 // includes all the other Perms.
 func (p Perms) Include(other Perms) bool {
-	return other != None && p&other == other
+	return p&other == other
 }
 
 // String implements the fmt.Stringer interface.
 func (p Perms) String() string {
-	var sb strings.Builder
-
-	for mask := Read; mask != 0; mask >>= 1 {
-		switch p & mask {
-		case Read:
-			sb.WriteString("read,")
-		case Write:
-			sb.WriteString("write,")
-		}
+	switch p {
+	case Read:
+		return "read"
+	case Write:
+		return "write"
+	case Read | Write:
+		return "read,write"
+	default:
+		return "none"
 	}
-
-	if hi := sb.Len(); hi > 0 {
-		return sb.String()[:hi-1]
-	}
-
-	return "none"
 }