… cmd/frontend/authz/perms_test.go
… cmd/frontend/db/repos_db_test.go
… cmd/frontend/db/repos_perm_filter_test.go
… cmd/frontend/db/repos_perm_test.go
… enterprise/cmd/frontend/internal/authz/authz_test.go
… enterprise/cmd/frontend/internal/authz/bitbucketserver/provider.go
… enterprise/cmd/frontend/internal/authz/bitbucketserver/provider_test.go
… enterprise/cmd/frontend/internal/authz/bitbucketserver/store.go
… enterprise/cmd/frontend/internal/authz/github/github.go
… enterprise/cmd/frontend/internal/authz/github/github_cache_test.go
… enterprise/cmd/frontend/internal/authz/github/github_test.go
… enterprise/cmd/frontend/internal/authz/gitlab/common_test.go
… enterprise/cmd/frontend/internal/authz/gitlab/gitlab_oauth.go
… enterprise/cmd/frontend/internal/authz/gitlab/gitlab_oauth_test.go
… enterprise/cmd/frontend/internal/authz/gitlab/gitlab_sudo.go
… enterprise/cmd/frontend/internal/authz/gitlab/gitlab_sudo_test.go
… go.mod