@tsenart 2019-07-03
Current state of this benchmark:

```
name             time/op
_authzFilter-16  75.3ms ± 1%

name             alloc/op
_authzFilter-16  55.3MB ± 0%

name             allocs/op
_authzFilter-16   65.8k ± 0%
```