@tsenart 2019-07-05
Got another small boost with the last commit :)

```
name             old time/op    new time/op    delta
_authzFilter-16    2.38ms ± 4%    1.94ms ± 4%  -18.52%  (p=0.000 n=15+13)

name             old alloc/op   new alloc/op   delta
_authzFilter-16    3.23MB ± 0%    2.98MB ± 0%   -7.61%  (p=0.000 n=15+10)

name             old allocs/op  new allocs/op  delta
_authzFilter-16      97.0 ± 0%      96.0 ± 0%   -1.03%  (p=0.000 n=15+15)
```