@tsenart 2019-07-05
And another.

```
name             old time/op    new time/op    delta
_authzFilter-16    1.99ms ± 4%    1.88ms ± 6%  -5.67%  (p=0.000 n=14+15)

name             old alloc/op   new alloc/op   delta
_authzFilter-16    2.98MB ± 0%    2.98MB ± 0%    ~     (p=0.680 n=15+14)

name             old allocs/op  new allocs/op  delta
_authzFilter-16      96.0 ± 0%      96.0 ± 0%    ~     (all equal)
```