@tsenart 2019-07-05
And at last, the previous commit brings it further down.

```
name             old time/op    new time/op    delta
_authzFilter-16    1.88ms ± 6%    1.26ms ± 9%  -33.02%  (p=0.000 n=15+15)

name             old alloc/op   new alloc/op   delta
_authzFilter-16    2.98MB ± 0%    0.86MB ± 2%  -71.24%  (p=0.000 n=14+15)

name             old allocs/op  new allocs/op  delta
_authzFilter-16      96.0 ± 0%      51.0 ± 0%  -46.88%  (p=0.000 n=15+15)
```