@tsenart 2019-07-06
Some more improvements

```
name             old time/op    new time/op    delta
_authzFilter-16    1.21ms ± 8%    1.05ms ± 1%  -13.09%  (p=0.000 n=15+15)

name             old alloc/op   new alloc/op   delta
_authzFilter-16     884kB ± 1%     537kB ± 0%  -39.29%  (p=0.000 n=15+14)

name             old allocs/op  new allocs/op  delta
_authzFilter-16      51.0 ± 0%      39.0 ± 0%  -23.53%  (p=0.000 n=15+15)
```